[improvement][headless]Declare QueryParser and QueryPlanner as interface instead of class.
