+++ headless/core/src/main/java/com/tencent/supersonic/headless/core/planner/QueryOptimizer.java
@@ -1,11 +1,10 @@
 package com.tencent.supersonic.headless.core.planner;
 
-import com.tencent.supersonic.headless.api.request.QueryStructReq;
 import com.tencent.supersonic.headless.core.pojo.QueryStatement;
 
 /**
- * the interface that rewrites QueryStatement with some optimization rules
+ * A query optimizer rewrites QueryStatement with a set of optimization rules
  */
 public interface QueryOptimizer {
-    void rewrite(QueryStructReq queryStructCmd, QueryStatement queryStatement);
+    void rewrite(QueryStatement queryStatement);
 }
